🚚 글로벌 타입 파일 통일
