--- extension/src/@types/Problem.ts
@@ -1,12 +1,0 @@
-export type ProblemResponse = {
-  problemId: number;
-  tier: number;
-  problemTitle: string;
-  score?: number;
-  teamName?: string;
-};
-
-export type ProblemRequest = {
-  userId: number;
-  problemNumber: number;
-};
