--- extension/src/@types/Ranking.ts
@@ -1,5 +1,0 @@
-export type Ranking = {
-  teamName: string;
-  bojId: string;
-  score: number;
-};
