--- extension/src/@types/Solving.ts
@@ -1,12 +1,0 @@
-export type SolvingProblemRequest = {
-  userId: number;
-  problemNumber: number;
-};
-
-export type SolvingProblemResponse = {
-  problemId: number;
-  tier: number;
-  problemTitle: string;
-  score: number;
-  teamName: string;
-};
