--- extension/src/@types/UnsolvedUser.ts
@@ -1,5 +1,0 @@
-export type UnsolvedUser = {
-  id: number;
-  bojId: string;
-  solvingCount: number;
-};
