+++ extension/src/contentScript/types/index.ts
